+++ pystache/loader.py
@@ -10,42 +10,45 @@ import os
 
 class Loader(object):
 
-    template_extension = 'mustache'
-    template_path = '.'
-    template_encoding = None
+    def __init__(self, search_dirs=None, encoding=None, extension=None):
+        """
+        Construct a template loader.
 
-    def load_template(self, template_name, template_dirs=None, encoding=None, extension=None):
+        Arguments:
+
+          search_dirs: the directories in which to search for templates.
+            Defaults to the current working directory.
+
+        """
+        if extension is None:
+            extension = 'mustache'
+        if search_dirs is None:
+            search_dirs = os.curdir  # i.e. "."
+
+        if isinstance(search_dirs, basestring):
+            search_dirs = [search_dirs]
+
+        self.search_dirs = search_dirs
+        self.template_encoding = encoding
+        self.template_extension = extension
+
+    def load_template(self, template_name):
         """
         Find and load the given template, and return it as a string.
 
         Raises an IOError if the template cannot be found.
 
         """
-        if None == template_dirs:
-            template_dirs = self.template_path
-
-        if encoding is not None:
-            self.template_encoding = encoding
-
-        if extension is not None:
-            self.template_extension = extension
-
+        search_dirs = self.search_dirs
         file_name = template_name + '.' + self.template_extension
 
-        # Given a single directory, we'll load from it.
-        if isinstance(template_dirs, basestring):
-            file_path = os.path.join(template_dirs, file_name)
-
-            return self._load_template_file(file_path)
-
-        # Given a list of directories, we'll check each for our file.
-        for path in template_dirs:
-            file_path = os.path.join(path, file_name)
+        for dir_path in search_dirs:
+            file_path = os.path.join(dir_path, file_name)
             if os.path.exists(file_path):
                 return self._load_template_file(file_path)
 
         # TODO: we should probably raise an exception of our own type.
-        raise IOError('"%s" not found in "%s"' % (template_name, ':'.join(template_dirs),))
+        raise IOError('"%s" not found in "%s"' % (template_name, ':'.join(search_dirs),))
 
     def _load_template_file(self, file_path):
         """